Add test for passing non-trivial structures as varargs (#9232)
Fixes #9042.

Renamed old test_struct_varargs to test_trivial_struct_varargs since
they only tested single member structs, which are handled differently.
