+++ tests/core/test_struct_varargs.c
@@ -7,72 +7,34 @@
 
 #include <stdio.h>
 #include <stdarg.h>
-#include <stdlib.h>
 
-struct A {
+typedef struct {
   int x;
-};
+  int y;
+} coords;
 
-struct B {
-  double x;
-};
+typedef struct {
+  int r;
+  int g;
+  int b;
+} color;
 
-void foo(int unused, ...)
-{
-  va_list vl;
-  va_start(vl, unused);
-  struct A a = va_arg(vl, struct A);
-  struct B b = va_arg(vl, struct B);
-  va_end(vl);
-
-  printf("%d\n", a.x);
-  printf("%f\n", b.x);
-}
-
-void a() {
-  struct A a = {
-    .x = 42,
-  };
-  struct B b = {
-    .x = 42.314,
-  };
-  foo(0, a, b);
-}
-
-struct tiny
-{
-  short c;
-};
-
-void f (int n, ...)
-{
-  struct tiny x;
-  int i;
+void fun(int a, ...) {
+  (void)a;
   va_list ap;
-  va_start (ap,n);  
-  for (i = 0; i < n; i++)
-  {
-    x = va_arg (ap,struct tiny);
-    printf("%d : %d\n", i, x.c);
-    if (x.c != i + 10) abort();
-  }
-  va_end (ap);
+  va_start(ap, a);
+
+  coords var1 = va_arg(ap, coords);
+  color var2 = va_arg(ap, color);
+  printf("va_arg coords: %d and %d\n", var1.x, var1.y);
+  printf("va_arg color: %d, %d, %d\n", var2.r, var2.g, var2.b);
+
+  va_end(ap);
 }
 
-void b ()
-{
-  struct tiny x[3];
-  struct tiny y;
-  printf("sizeof tiny: %d (3 of them: %d)\n", sizeof(y), sizeof(x));
-  x[0].c = 10;
-  x[1].c = 11;
-  x[2].c = 12;
-  f (3, x[0], x[1], x[2]);
+int main(void) {
+  coords val1 = { .x = 42, .y = 21 };
+  color val2 = { .r = 37, .g = 19, .b = 253 };
+  fun(0, val1, val2);
+  return 0;
 }
-
-int main() {
-  a();
-  b();
-  printf("ok.\n");
-}
-
