+++ tests/core/test_trivial_struct_varargs.out
@@ -1,0 +1,7 @@
+42
+42.314000
+sizeof tiny: 2 (3 of them: 6)
+0 : 10
+1 : 11
+2 : 12
+ok.
